Updates to handling of C0 and D0
Removed setter function for initial condition on cases and deaths. We now compute/set these in the initial conditions function but the user shouldn't touch these. We only count deaths from t=0. Changed names of slots and associated functions accordingly

Changed name of parameter a to parameter m, because I believe mu is the conventional name for the death rate.

Changed name of simulate_SEIR to simulate_SEIDR to stick with naming convention across models.

Do we need to reformat the output?