Fighting setter methods
Changed the argument names in the setter function. Apparantly the second argument must be named value, and it must be one argument. Rewritten these methods to work for this.

For some reason the test code at the bottom os SEIDR.R works fine but trying to run the exact same lines of code in the vignette still doesn't. No clue why, will try again later.