Implemented second version urban vs. rural
Created a class for the urban vs. rural model where migration between communities is allowed, but only infection within a community is possible. Also edited the vignette to show a comparison of the two models.

It's all a bit rough and ready at the moment cause I just wanted to make sure the models worked and to see how they compare but the there will most certainly be errors in the documentation and the vignette. Also, there are no tests yet for either of these new urban vs. rural classes.